+++ support/support_nix.go
@@ -460,9 +460,10 @@ func (i *Info) printDependencies() {
 	fmtutil.Separator(false, "DEPENDENCIES")
 
 	for _, dep := range i.Deps {
-		if dep.Extra == "" {
+		switch dep.Extra {
+		case "":
 			fmtc.Printf(" {s}%8s{!}  %s\n", dep.Version, dep.Path)
-		} else {
+		default:
 			fmtc.Printf(" {s}%8s{!}  %s {s-}(%s){!}\n", dep.Version, dep.Path, dep.Extra)
 		}
 	}